feat(parser): add support for reference types
