+++ src/lexer/token_kind.rs
@@ -129,4 +129,8 @@ pub enum TokenKind {
     LessOrEqual {
         position: Position,
     },
+    #[terminal]
+    Ampersand {
+        position: Position,
+    },
 }
